requirements.txt
numpy==1.11.2
scikit-learn==0.18.1
Keras==1.1.0
matplotlib==1.5.1
Pillow==3.1.2
